Convert pixel writing routines to write four pixels at a time.
git-svn-id: file:///var/lib/svn/backer/trunk@615 4b75c72c-9afd-0310-a208-b187eb4cea4a

+++ codecs/bkr_video_out.h
@@ -53,7 +53,7 @@ typedef struct {
 		gint interlace;
 		gint width;
 		gint height;
-		guint8 *(*pixel_func)(guint8 *, guint8);
+		guint32 *(*pixel_func)(guint32 *, guint32);
 	} format;
 } BkrVideoOut;
